Various MS auth improvements - Add --offline - Add --profile
